comment handling for 'super' and 'zsuper'
Summary:
This work ensures inline comments are attached after the 'super' and
'zsuper' nodes.
