+++ test/js/super.test.js
@@ -11,4 +11,23 @@ describe("super", () => {
 
   test("multiple args, with parens", () =>
     expect("super(1, 2)").toMatchFormat());
+
+  describe("with comment", () => {
+    test("bare", () => expect("super # comment").toMatchFormat());
+
+    test("empty parens", () =>
+      expect("super() # comment").toMatchFormat());
+
+    test("one arg, no parens", () =>
+    expect("super 1 # comment").toMatchFormat());
+
+    test("one arg, with parens", () =>
+      expect("super(1) # comment").toMatchFormat());
+
+    test("multiple args, no parens", () =>
+      expect("super 1, 2 # comment").toMatchFormat());
+
+    test("multiple args, with parens", () =>
+      expect("super(1, 2) # comment").toMatchFormat());
+  })
 });
